move proxy methods to module proxy
