+++ events/messageCreate.js
@@ -1,5 +1,4 @@
 module.exports = async (msg,bot) => {
 	let ctx = await bot.getMessageContext(msg);
-	if (ctx.done) return;
-	if (await bot.cmd(ctx) && msg.channel.guild) bot.proxy(ctx);
+	if (!ctx.done && await bot.cmd(ctx) && msg.channel.guild && !(await bot.db.isBlacklisted(msg.channel.guild.id, msg.channel.id, true))) bot.proxy.executeProxy(ctx);
 };
